# docs/aidlc/05-api-specification/Fantasy EPL League Manager — OpenAPI Specification v1.0.yaml
openapi: 3.0.3
info:
  title: Fantasy EPL League Manager API
  version: "1.0"
  description: |
    # Fantasy EPL League Manager
    ## OpenAPI Specification — Version 1.0

    **Document Status:** Baseline API Specification

    **Inputs:**
    - Architecture and Domain Model v1.15 (`Fantasy EPL League Manager — Architecture and Domain Model v1.15.md`) — Section 9's endpoint sketches (§9.2), the aggregate/entity model (§6), logical data architecture (§8), and cross-cutting API conventions (§9.3) that this document turns into concrete request/response DTOs and validation rules.
    - Business Requirements Document v1.17 (`Fantasy EPL League Manager — Business Requirements Document v1.17.md`) — authoritative business rules, cited throughout by `BR-###`.
    - Epic and Feature Backlog v1.10 (`Fantasy EPL League Manager — Epic and Feature Backlog v1.10.md`) — feature IDs (`F-###.#`) cited on each operation group to show which feature it serves and in what build-sequence position (Backlog §5).

    **Purpose:** Turn Architecture §9's illustrative endpoint list into a complete, implementation-ready OpenAPI contract — every resource, request/response schema, validation rule, and error shape needed to start writing controllers and client code. This is the artifact Architecture §16 lists as recommended-next-artifact #3.

...

...

    ## Conventions (Architecture §9.1/§9.3)

    - REST over HTTPS only; JSON request/response bodies.
...

    ## Version History

    ### Version 1.0

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /leagues/{leagueId}/seasons/{seasonId}/drafts:
    parameters:
      - $ref: '#/components/parameters/leagueId'
      - $ref: '#/components/parameters/seasonId'
    post:
      tags: [Draft]
      operationId: createDraft
      summary: Schedule a Draft — Initial, Secondary, or Replacement (F-005.1/F-006.1, BR-053–BR-055, BR-069, BR-281)
      description: |
        For `draftType: Secondary`, the proposed `startTime` defaults to the transfer-window close date plus
        `SeasonConfiguration.SecondaryDraftSchedulingOffsetDays`, advanced past any day with a scheduled EPL fixture
        (BR-069/BR-281); the League Administrator may override it in the request.
      x-authorization: Caller must be the League Administrator of leagueId.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [draftType]
              properties:
                draftType:
                  $ref: '#/components/schemas/DraftType'
                startTime: { type: string, format: date-time }
      responses:
        '201':
          description: Draft scheduled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Draft'
        '409':
          description: An Initial Draft was requested with fewer than two FantasyTeams participating (BR-302).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    get:
      tags: [Draft]
      operationId: listDrafts
      summary: List Drafts for a Season
      x-authorization: Caller must be an active member of leagueId.
      parameters:
        - name: draftType
          in: query
          schema:
            $ref: '#/components/schemas/DraftType'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Draft'
  /drafts/{draftId}:
    parameters:
      - $ref: '#/components/parameters/draftId'
    get:
      tags: [Draft]
      operationId: getDraft
      summary: Get Draft state — order, current round/pick, timer deadline, pending makeup picks (F-005.5, BR-204, BR-205, BR-207, BR-208, BR-324)
      x-authorization: Caller must be an active member of the Draft's League.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Draft'
  /drafts/{draftId}/start:
    post:
      tags: [Draft]
      operationId: startDraft
      summary: Start a scheduled Draft (League Administrator only)
      x-authorization: Caller must be the League Administrator of the Draft's League.
      parameters:
        - $ref: '#/components/parameters/draftId'
      responses:
        '200':
          description: Draft is now InProgress.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Draft'
        '409':
          description: Fewer than two FantasyTeams participate (BR-302, Initial Draft only).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /drafts/{draftId}/pause:
    post:
      tags: [Draft]
      operationId: pauseDraft
      summary: Pause an in-progress Draft (League Administrator only, BR-025)
      x-authorization: Caller must be the League Administrator of the Draft's League.
      parameters:
        - $ref: '#/components/parameters/draftId'
      responses:
        '200':
          description: Paused.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Draft'
...
  /leagues/{leagueId}/audit:
    get:
      tags: [Administration]
      operationId: getAuditLog
      summary: Administrative audit log, filterable (F-011.1, BR-177–BR-182, BR-321, BR-322)
      description: A row with no resolvable FantasyTeamId (e.g., ActionType = ConfigurationChanged) is surfaced under the pseudo-value `"__league_settings__"` for the fantasyTeamId filter, per Architecture §6.8.
      x-authorization: Caller must be the League Administrator of leagueId.
      parameters:
        - $ref: '#/components/parameters/leagueId'
        - name: actionType
          in: query
          schema:
            $ref: '#/components/schemas/AdminActionType'
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
        - name: fantasyTeamId
          in: query
          schema: { type: string }
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/cursor'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdministrativeActionPage'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    leagueId:
      name: leagueId
      in: path
      required: true
      schema: { type: string, format: uuid }
    seasonId:
      name: seasonId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    draftId:
      name: draftId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    cursor:
      name: cursor
      in: query
      schema: { type: string }
  responses:
...
    Forbidden:
      description: Authenticated, but the caller fails the operation's object-level authorization check (BR-163, BR-254).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
  schemas:
    ProblemDetails:
      type: object
      description: RFC 7807-style error body (Architecture §9.3), extended with a stable machine-readable errorCode.
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        errorCode:
          type: string
          example: player_already_owned
        correlationId: { type: string, description: Propagated into logs (BR-241). }
...
    DraftType:
      type: string
      enum: [Initial, Secondary, Replacement]
    DraftStatus:
      type: string
      enum: [Scheduled, InProgress, Paused, Completed]
    Draft:
      type: object
      properties:
        draftId: { type: string, format: uuid }
        seasonId: { type: string, format: uuid }
        draftType:
          $ref: '#/components/schemas/DraftType'
        status:
          $ref: '#/components/schemas/DraftStatus'
        draftOrder:
          type: array
          items: { type: string, format: uuid }
          description: FantasyTeamId sequence — randomized for Initial (BR-054), standings-derived for Secondary (BR-056).
        currentRound: { type: integer }
        currentPickIndex: { type: integer }
        timerSeconds: { type: integer, description: "BR-057" }
        currentPickDeadline: { type: string, format: date-time, nullable: true }
        pendingMakeupPicks:
          type: array
          items: { type: string, format: uuid }
          description: FantasyTeamIds queued for a makeup pick after the final round (BR-282, BR-324).
...
    AdminActionType:
      type: string
      enum:
        - RosterCorrection
        - ScoreOverride
        - ScoreOverrideUndo
        - ReplacementEligibilityGranted
        - SeasonEndingInjuryDeclared
        - DraftTimerExtended
        - ConfigurationChanged
        - Other
    AdministrativeAction:
      type: object
      properties:
        actionId: { type: string, format: uuid }
        leagueId: { type: string, format: uuid }
        actingMembershipId: { type: string, format: uuid, nullable: true, description: "Null means system-generated (e.g. automatic EPL-exit eligibility grant, BR-308) — render as \"System\"." }
        actionType:
          $ref: '#/components/schemas/AdminActionType'
        targetEntityType: { type: string }
        targetEntityId: { type: string, format: uuid }
        beforeState: { type: object, additionalProperties: true }
        afterState: { type: object, additionalProperties: true }
        reason: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    AdministrativeActionPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AdministrativeAction'
        nextCursor: { type: string, nullable: true }
...